🧹 updates folder structure for organization (#8)
* 🧹 updates folder structure for organization

* 🧹 removes clg

* 🪲 updates import
