+++ app/_components/common/Overlay/Overlay.tsx
@@ -1,0 +1,5 @@
+export default function Overlay() {
+  return (
+    <div className='absolute bg-blue-400 h-full opacity-20 overscroll-none top-0 w-full z-0' />
+  );
+}
