+++ app/_components/common/index.ts
@@ -1,0 +1,5 @@
+import Header from '@components/common/Header/Header';
+import Overlay from '@components/common/Overlay/Overlay';
+import SearchBar from '@components/common/SearchBar/SearchBar';
+
+export { Header, Overlay, SearchBar };
