+++ app/_components/nba-everything/NBAEverythingSeasonAverages/components/NBAEverythingSeasonAverage/NBAEverythingSeasonAverage.tsx
