+++ app/_components/nba-everything/NBAEverythingSeasonAverages/components/index.ts
@@ -1,0 +1,3 @@
+import NBASeasonAverage from '@components/nba-everything/NBAEverythingSeasonAverages/components/NBAEverythingSeasonAverage/NBAEverythingSeasonAverage';
+
+export { NBASeasonAverage };
