+++ app/_components/nba-everything/NBAEverythingTeamGames/components/index.ts
@@ -1,0 +1,3 @@
+import NBAEverythingTeamGame from '@components/nba-everything/NBAEverythingTeamGames/components/NBAEverythingTeamGame/NBAEverythingTeamGame';
+
+export { NBAEverythingTeamGame };
