+++ app/_components/nba-everything/index.ts
@@ -1,0 +1,17 @@
+import NBAEverything from '@components/nba-everything/NBAEverything/NBAEverything';
+import NBAEverythingSeasonAverages from '@components/nba-everything/NBAEverythingSeasonAverages/NBAEverythingSeasonAverages';
+import NBAEverythingSeasonSearch from '@components/nba-everything/NBAEverythingSeasonSearch/NBAEverythingSeasonSearch';
+import NBAEverythingStat from '@components/nba-everything/NBAEverythingStat/NBAEverythingStat';
+import NBAEverythingTeamGames from '@components/nba-everything/NBAEverythingTeamGames/NBAEverythingTeamGames';
+import NBAEverythingTeamLogo from '@components/nba-everything/NBAEverythingTeamLogo/NBAEverythingTeamLogo';
+import NBAEverythingTeamSearch from '@components/nba-everything/NBAEverythingTeamSearch/NBAEverythingTeamSearch';
+
+export default NBAEverything;
+export {
+  NBAEverythingSeasonAverages,
+  NBAEverythingSeasonSearch,
+  NBAEverythingStat,
+  NBAEverythingTeamGames,
+  NBAEverythingTeamLogo,
+  NBAEverythingTeamSearch,
+};
