+++ app/_styles/globals.css
@@ -1,0 +1,6 @@
+@import url('buttons/buttons.css');
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+
