+++ app/_types/NBAGame/NBAGame.ts
@@ -1,0 +1,21 @@
+import { NBATeam } from '@types';
+
+type NBAGame = {
+  id: number;
+  date: string;
+  datetime: string;
+  home_team_score: number;
+  home_team?: NBATeam;
+  home_team_id?: NBATeam;
+  period: number;
+  postseason: boolean;
+  season: number;
+  status: string;
+  time: string;
+  visitor_team?: NBATeam;
+  visitor_team_id?: NBATeam;
+  visitor_team_score: number;
+  win?: boolean;
+};
+
+export default NBAGame;
