+++ app/_types/NBAGameBoxScore/NBAGameBoxScoreShort.ts
