+++ app/_types/NBATeam/NBATeam.ts
@@ -1,0 +1,11 @@
+type NBATeam = {
+  id: number;
+  conference: string;
+  division: string;
+  city: string;
+  name: string;
+  full_name: string;
+  abbreviation: string;
+};
+
+export default NBATeam;
