--- app/components/index.ts
@@ -1,5 +1,0 @@
-import Header from '@app/components/Header/Header';
-import Overlay from '@app/components/Overlay/Overlay';
-import SearchBar from '@app/components/SearchBar/SearchBar';
-
-export { Header, Overlay, SearchBar };
