+++ app/game/[gameId]/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 import NBAEverythingGame from '@app/game/[gameId]/components';
-import { Overlay } from '@app/components';
+import { Overlay } from '@components/common';
 
 export default function GameIndex() {
   return (
