--- app/nba-everything/components/NBAEverythingSeasonAverages/components/index.ts
@@ -1,3 +1,0 @@
-import NBASeasonAverage from '@nba-everything/components/NBAEverythingSeasonAverages/components/NBAEverythingSeasonAverage/NBAEverythingSeasonAverage';
-
-export { NBASeasonAverage };
