--- app/nba-everything/components/NBAEverythingTeamGames/components/index.ts
@@ -1,3 +1,0 @@
-import NBAEverythingTeamGame from '@nba-everything/components/NBAEverythingTeamGames/components/NBAEverythingTeamGame/NBAEverythingTeamGame';
-
-export { NBAEverythingTeamGame };
