--- app/nba-everything/components/index.ts
@@ -1,15 +1,0 @@
-import NBAEverythingSeasonAverages from '@nba-everything/components/NBAEverythingSeasonAverages/NBAEverythingSeasonAverages';
-import NBAEverythingSeasonSearch from '@nba-everything/components/NBAEverythingSeasonSearch/NBAEverythingSeasonSearch';
-import NBAEverythingStat from '@nba-everything/components/NBAEverythingStat/NBAEverythingStat';
-import NBAEverythingTeamGames from '@nba-everything/components/NBAEverythingTeamGames/NBAEverythingTeamGames';
-import NBAEverythingTeamLogo from '@nba-everything/components/NBAEverythingTeamLogo/NBAEverythingTeamLogo';
-import NBAEverythingTeamSearch from '@nba-everything/components/NBAEverythingTeamSearch/NBAEverythingTeamSearch';
-
-export {
-  NBAEverythingSeasonAverages,
-  NBAEverythingSeasonSearch,
-  NBAEverythingStat,
-  NBAEverythingTeamGames,
-  NBAEverythingTeamLogo,
-  NBAEverythingTeamSearch,
-};
