--- app/nba-everything/layout/NBAEverythingLayout.tsx
@@ -1,12 +1,0 @@
-'use client';
-
-import queryClient from '@client/queryClient';
-import { QueryClientProvider as NBAEverythingClientProvider } from '@tanstack/react-query';
-
-type NBAEverythingLayoutProps = {
-  children: React.ReactNode;
-};
-
-export default function NBAEverythingLayout({ children }: NBAEverythingLayoutProps) {
-  return <NBAEverythingClientProvider client={queryClient}>{children}</NBAEverythingClientProvider>;
-}
