--- store/index.ts
@@ -1,3 +1,0 @@
-import useNBAEverythingStore from '@store/useNBAEverything/useNBAEverything';
-
-export { useNBAEverythingStore };
