--- styles/globals.css
@@ -1,7 +1,0 @@
-@import url('buttons/buttons.css');
-
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-
