+++ tsconfig.json
@@ -21,11 +21,14 @@
     "baseUrl": "./",
     "paths": {
       "@*": ["*"],
-      "@client/*": ["client/*"],
-      "@components/*": ["app/components/*"],
-      "@nba-everything/*": ["app/nba-everything/*"],
+      "@api/*": ["app/_api/*"],
+      "@client/*": ["app/_client/*"],
+      "@components/*": ["app/_components/*"],
+      "@store": ["app/_store"],
+      "@styles/*" : ["app/_styles/*"],
+      "@types": ["app/_types"],
       "@public/*": ["public/*"],
-      "@types/*": ["types/*"],
+      "@queryClient": ["client/queryClient"],
     },
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
