--- types/NBAGame/NBAGame.ts
@@ -1,19 +1,0 @@
-import { NBATeam } from '@types';
-
-export type NBAGame = {
-  id: number;
-  date: string;
-  datetime: string;
-  home_team_score: number;
-  home_team?: NBATeam;
-  home_team_id?: NBATeam;
-  period: number;
-  postseason: boolean;
-  season: number;
-  status: string;
-  time: string;
-  visitor_team?: NBATeam;
-  visitor_team_id?: NBATeam;
-  visitor_team_score: number;
-  win?: boolean;
-};
