--- types/NBATeam/NBATeam.ts
@@ -1,9 +1,0 @@
-export type NBATeam = {
-  id: number;
-  conference: string;
-  division: string;
-  city: string;
-  name: string;
-  full_name: string;
-  abbreviation: string;
-};
