--- types/index.ts
@@ -1,17 +1,0 @@
-import { NBATeam } from './NBATeam/NBATeam';
-import { NBATeamStats } from './NBATeamStats/NBATeamStats';
-import { NBAGame } from './NBAGame/NBAGame';
-import NBAGameBoxScore from './NBAGameBoxScore/NBAGameBoxScore';
-import NBAGameBoxScoreShort from './NBAGameBoxScore/NBAGameBoxScoreShort';
-import NBAGameStat from './NBAGameStat/NBAGameStat';
-import NBAGameStats from './NBAGameStats/NBAGameStats';
-
-export type {
-  NBAGame,
-  NBAGameBoxScore,
-  NBAGameBoxScoreShort,
-  NBAGameStat,
-  NBAGameStats,
-  NBATeam,
-  NBATeamStats,
-};
